refactor: 마이 페이지  타입 폴더 구조 변경
