+++ src/app/my/_hooks/queries/useReviewLabelList.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 
-import { reviewLabelList } from "../../../api/mypage";
+import { reviewLabelList } from "../../_api/mypage";
 
 const useReviewLabelList = (userId: number) => {
   return useQuery({
